BUG#: 7058 TITLE: Compiler warnings in CMPI code. DESCRIPTION:

+++ src/Providers/TestProviders/CMPI/TestIndications/cmpiTestIndicationProvider.c
@@ -88,7 +88,7 @@ gettimeofday (struct timeval *t, void *timezone)
 {
     struct _timeb timebuffer;
     _ftime (&timebuffer);
-    t->tv_sec = timebuffer.time;
+    t->tv_sec = (long)timebuffer.time;
     t->tv_usec = 1000 * timebuffer.millitm;
 }
 #endif
